eval changed to json and save and load given own file
