+++ lompe/__init__.py
@@ -6,3 +6,4 @@ from lompe.utils import sunlight
 from lompe.utils import conductance
 from lompe.model import visualization
 from lompe.secsy import cubedsphere as cs
+from lompe.utils import load_model, load_grid, save_model
